+++ Locales/deDE.lua
@@ -1,3 +1,4 @@
-if GetLocale() ~= "deDE" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "deDE")
+if not L then return end
 
 L["Corruption"] = "Verderbnis"