+++ Locales/enUS.lua
@@ -1,2 +1,15 @@
-L = {}
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "enUS", true)
+if not L then return end
+
 L["Corruption"] = "Corruption"
+L["I"] = "I"
+L["II"] = "II"
+L["III"] = "III"
+
+--config
+L["Append to corruption stat"] = "Append to corruption stat"
+L["Use the new style tooltip."] = "Use the new style tooltip."
+L["Show summary on the corruption tooltip"] = "Show summary on the corruption tooltip"
+L["List your corruptions in the eye tooltip in the character screen."] = "List your corruptions in the eye tooltip in the character screen."
+L["Display in English"] = "Display in English"
+L["Don't translate the corruption effect names."] = "Don't translate the corruption effect names."