+++ Locales/esES.lua
@@ -1,3 +1,4 @@
-if GetLocale() ~= "esES" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "esES")
+if not L then return end
 
 L["Corruption"] = "Corrupción"