+++ Locales/esMX.lua
@@ -1,3 +1,4 @@
-if GetLocale() ~= "esMX" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "esMX")
+if not L then return end
 
 L["Corruption"] = "Corrupción"