+++ Locales/frFR.lua
@@ -1,3 +1,4 @@
-if GetLocale() ~= "frFR" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "frFR")
+if not L then return end
 
 L["Corruption"] = "Corruption"