+++ Locales/ruRU.lua
@@ -1,3 +1,4 @@
-if GetLocale() ~= "ruRU" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "ruRU")
+if not L then return end
 
 L["Corruption"] = "порче"