+++ Locales/zhCN.lua
@@ -1,3 +1,7 @@
-if GetLocale() ~= "zhCN" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "znCN")
+if not L then return end
 
-L["Corruption"] = "腐蚀术"
+L["Corruption"] = "腐蚀"
+L["I"] = "1 级"
+L["II"] = "2 级"
+L["III"] = "3 级"