+++ Locales/zhTW.lua
@@ -1,3 +1,4 @@
-if GetLocale() ~= "zhTW" then return end
+local L = LibStub("AceLocale-3.0"):NewLocale("CorruptionTooltips", "zhTW")
+if not L then return end
 
 L["Corruption"] = "腐化"